fixed-nonexisting-fields-bug
