+++ src/main/java/com/internationalairport/airportmanagementsystem/daos/FlightCrewRepository.java
@@ -2,5 +2,5 @@ package com.internationalairport.airportmanagementsystem.daos;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FlightCrewRepository extends JpaRepository<FlightCrew, Integer> {
-}
+//public interface FlightCrewRepository extends JpaRepository<FlightCrew, Integer> {
+//}
