+++ src/main/java/com/internationalairport/airportmanagementsystem/entities/Employee.java
@@ -25,17 +25,16 @@ public class Employee {
     @Column(name = "contact_info")
     private String contactInfo;
 
-    @ManyToOne(cascade = {CascadeType.DETACH, CascadeType.MERGE,
-            CascadeType.PERSIST, CascadeType.REFRESH})
-    @JoinColumn(name = "airport_id")
-    @JsonManagedReference
-    private Airport airport;
-
-    @OneToMany(mappedBy = "employee",
-            cascade = {CascadeType.DETACH, CascadeType.MERGE,
-            CascadeType.PERSIST, CascadeType.REFRESH})
+    @ManyToMany(
+            cascade = {CascadeType.PERSIST, CascadeType.MERGE,
+                    CascadeType.DETACH, CascadeType.REFRESH})
+    @JoinTable(
+            name = "flight_crews",
+            joinColumns = @JoinColumn(name = "employee_id"),
+            inverseJoinColumns = @JoinColumn(name = "flight_id")
+    )
     @JsonBackReference
-    private List<FlightCrew> flightCrews;
+    private List<Flight> flights;
 
     // Constructors, getters, and setters
     public Employee() {
@@ -80,14 +79,6 @@ public class Employee {
         this.contactInfo = contactInfo;
     }
 
-    public Airport getAirport() {
-        return airport;
-    }
-
-    public void setAirport(Airport airport) {
-        this.airport = airport;
-    }
-
     @Override
     public String toString() {
         return "Employee{" +
@@ -98,22 +89,20 @@ public class Employee {
                 '}';
     }
 
-    public List<FlightCrew> getFlightCrews() {
-        return flightCrews;
+    public List<Flight> getFlights() {
+        return flights;
     }
 
-    public void setFlightCrews(List<FlightCrew> flightCrews) {
-        this.flightCrews = flightCrews;
+    public void setFlights(List<Flight> flights) {
+        this.flights = flights;
     }
 
     // define convenient methods
-    public void addFlightCrew(FlightCrew tempFlightCrew) {
-        if( flightCrews == null) {
-            flightCrews = new ArrayList<>();
+    public void addFlight(Flight tempFlight) {
+        if( flights == null) {
+            flights = new ArrayList<>();
         }
 
-        flightCrews.add(tempFlightCrew);
-
-        tempFlightCrew.setEmployee(this);
+        flights.add(tempFlight);
     }
 }
