+++ src/main/java/com/internationalairport/airportmanagementsystem/entities/FlightCrew.java
@@ -1,0 +1,74 @@
+package com.internationalairport.airportmanagementsystem.entities;
+
+import jakarta.persistence.*;
+
+//@Entity
+//@Table(name = "flight_crews")
+public class FlightCrew {
+//    @Id
+//    @GeneratedValue(strategy = GenerationType.IDENTITY)
+//    @Column(name = "crew_id")
+//    private Integer crewId;
+//
+//    @Column(name = "role")
+//    private String role;
+//
+//    @ManyToOne(cascade = {CascadeType.DETACH, CascadeType.MERGE,
+//            CascadeType.PERSIST, CascadeType.REFRESH})
+//    @JoinColumn(name = "flight_id")
+//    private Flight flight;
+//
+//    @ManyToOne(cascade = {CascadeType.DETACH, CascadeType.MERGE,
+//            CascadeType.PERSIST, CascadeType.REFRESH})
+//    @JoinColumn(name = "employee_id")
+//    private Employee employee;
+//
+//    // Constructors
+//    public FlightCrew() {
+//    }
+//
+//    public FlightCrew(String role) {
+//        this.role = role;
+//    }
+//
+//    // Getters and Setters
+//    public Integer getCrewId() {
+//        return crewId;
+//    }
+//
+//    public void setCrewId(Integer crewId) {
+//        this.crewId = crewId;
+//    }
+//
+//    public String getRole() {
+//        return role;
+//    }
+//
+//    public void setRole(String role) {
+//        this.role = role;
+//    }
+//
+//    public Flight getFlight() {
+//        return flight;
+//    }
+//
+//    public void setFlight(Flight flight) {
+//        this.flight = flight;
+//    }
+//
+//    public Employee getEmployee() {
+//        return employee;
+//    }
+//
+//    public void setEmployee(Employee employee) {
+//        this.employee = employee;
+//    }
+//
+//    @Override
+//    public String toString() {
+//        return "FlightCrew{" +
+//                "crewId=" + crewId +
+//                ", role='" + role + '\'' +
+//                '}';
+//    }
+}
