+++ src/main/java/com/internationalairport/airportmanagementsystem/service/interfaces/FlightCrewService.java
@@ -3,10 +3,10 @@ package com.internationalairport.airportmanagementsystem.service.interfaces;
 import java.util.List;
 
 public interface FlightCrewService {
-    FlightCrew save(FlightCrew flightCrew);
-    FlightCrew findById(Integer flightCrewId);
-    List<FlightCrew> findAll();
-    void deleteById(Integer flightCrewId);
-    String deleteAll();
+//    FlightCrew save(FlightCrew flightCrew);
+//    FlightCrew findById(Integer flightCrewId);
+//    List<FlightCrew> findAll();
+//    void deleteById(Integer flightCrewId);
+//    String deleteAll();
 }
